pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-restful>=0.3.10",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-cors>=5.0.1",
    "marshmallow-sqlalchemy>=1.4.2",
    "marshmallow>=4.0.0",
    "sqlalchemy>=2.0.40",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
    "openai>=1.93.0",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
]
